Add "forge" rule to rewrite A records
For some dns rewriting, it is not enough to replace queries with CNAMEs.
In particular, some dns clients will error if they request an A record
and get a CNAME as a response, and won't fall back to re-requesting a
CNAME.

This rule type adds an option that may be used to completely forge a
request for an A record, replacing the output IP addresses returned with
our own. There is also an option for doing an upstream DNS query and
using that response to replace the IP addresses in a target request.

Signed-off-by: Nick Stevens <nick@bitcurry.com>